+++ MIA.cpp
@@ -6,10 +6,10 @@
 #include <App.hpp>
 #include <Standard_Gen.hpp>
 
+#include <modules/EnumConversions.hpp>
+
 int main(int argc, char** argv)
 {
-	using Standard = mia::CppStandard;
-
 	// Setup the tool's arguments.
 	std::vector<std::string> filesToProcess, includeDirs;
 	std::string outputPattern;
@@ -17,13 +17,19 @@ int main(int argc, char** argv)
 	argumentum::argument_parser parser;
 	auto params = parser.params();
 
-	mia::AppConfig config;
+	mia::GeneratorConfig config;
 
 	parser.config().program(argv[0]).description("Meta Information Appender");
 
-	params.add_parameter(filesToProcess, "--files", "-f").minargs(1).metavar("<files>").help("Files to process").required(true);
-	params.add_parameter(includeDirs, "--includes", "-i").minargs(1).metavar("<include_dirs>").help("Include directories").required(false);
-	params.add_parameter(outputPattern, "--output", "-o").metavar("<output_pattern>").help("Output pattern for the output files. Defaults to \"{}.hpp\" where {} is a placeholder for the input file name").required(false).absent("{}.hpp").nargs(1);
+	params
+		.add_parameter(filesToProcess, "--files", "-f").minargs(1).metavar("<files>")
+		.help("Files to process").required(true);
+	params
+		.add_parameter(includeDirs, "--includes", "-i").minargs(1).metavar("<include_dirs>")
+		.help("Include directories").required(false);
+	params
+		.add_parameter(outputPattern, "--output", "-o").metavar("<output_pattern>")
+		.help("Output pattern for the output files. Defaults to \"{}.hpp\" where {} is a placeholder for the input file name").required(false).absent("{}.hpp").nargs(1);
 	
 	config.registerOptions(params);
 
@@ -32,6 +38,11 @@ int main(int argc, char** argv)
 		return 1;
 
 	mia::App app(std::move(filesToProcess), std::move(includeDirs), std::move(outputPattern), config);
+
+	app.registerModules({
+		std::make_shared<mia::modules::EnumConversionsModule>()
+	});
+
 	app.process();
 
 	return 0;
