+++ src/CMakeLists.txt
@@ -1,6 +1,11 @@
-target_sources(mia
-	PRIVATE
-		App.cpp
-		CodeGenerator.cpp
-		StringUtils.cpp
+target_sources(mia-core PRIVATE
+	StringUtils.cpp
+	CodeGenerator.cpp
+	Generator.cpp
 )
+
+target_sources(mia PRIVATE
+	App.cpp
+)
+
+add_subdirectory(modules)
