+++ src/modules/CMakeLists.txt
@@ -1,0 +1,3 @@
+target_sources(mia-core PRIVATE
+	EnumConversions.cpp
+)
